Improve: update evictable cursor after flush finished
Before this commit, the evcitable cursor is update when `flush` command
is sent, not when `flush` command is finished. Thus there is chance a
log is not in file or cache, if the wal-record is not written to file
but in the buffer.

Buffering wal record will be added in future.
This commit ensures a log is either in cache or in file.