refactor: cache should evict purged items even when cache is not full
The cache may be configured with a large capacity or max items, but purged items
will never be accessed again and should be removed, even when the cache's
max-item/capacity limit is not reached.

In this commit, when purging raft-logs, we also remove the corresponding items
from the cache when the purge-log command is synced.

Therefore the server memory usage will be minimized.